1. fixed Categorical init and gen function 2. added code::ClassConstructor 3. change all int to size_t whenever necessary 4. added translation for generic type for MapExpr 5. various bug fix in CPPTranslator

+++ src/absyn/MapExpr.cpp
@@ -21,11 +21,11 @@ size_t MapExpr::mapSize() {
   return args.size() >> (size_t) 1;
 }
 
-Expr* MapExpr::getFrom(int k) {
+Expr* MapExpr::getFrom(size_t k) {
   return args[k << 1];
 }
 
-Expr* MapExpr::getTo(int k) {
+Expr* MapExpr::getTo(size_t k) {
   return args[(k << 1) | 1];
 }
 
@@ -38,9 +38,9 @@ void MapExpr::addMap(Expr* from, Expr* to) {
 void MapExpr::print(FILE* file, int indent) {
   fprintf(file, "%*s(MapExpr:\n", indent, "");
   for (size_t i = 0; i < mapSize(); i++) {
-    fprintf(file, "%*s:from#%d\n", indent + 2, "", i);
+    fprintf(file, "%*s:from#%zu\n", indent + 2, "", i);
     getFrom(i)->print(file, indent + 4);
-    fprintf(file, "%*s:to#%d\n", indent + 2, "", i);
+    fprintf(file, "%*s:to#%zu\n", indent + 2, "", i);
     getTo(i)->print(file, indent + 4);
   }
   fprintf(file, "%*s)\n", indent, "");
